+++ python/GafferSceneUI/AttributeTweaksUI.py
@@ -98,6 +98,7 @@ Gaffer.Metadata.registerNode(
 		"tweaks.*" : [
 
 			"tweakPlugValueWidget:allowCreate", True,
+			"tweakPlugValueWidget:allowRemove", True,
 
 		],